refactor(model): model 내부에 클래스가 많아짐에 따라 세부 패키지 분리
- Lotto, LottoTicket, PurchaseAmount 클래스를 ticket 패키지 하위에 배치

- 이전에 작성한 테스트 코드인 PurchaseAmountTest 클래스도 패키지 경로에 맞게 이동
