+++ src/main/java/lotto/model/ticket/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import java.util.HashSet;
 import java.util.List;
