+++ src/main/java/lotto/model/ticket/LottoTickets.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import java.util.Collections;
 import java.util.List;
