+++ src/main/java/lotto/model/ticket/PurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import java.util.Objects;
 import lotto.exception.ErrorMessage;
